+++ CHANGES.md
@@ -1,4 +1,6 @@
 [unreleased]
+
+#### 0.4.0 / 2021-03-10
 * add data to `gu_api_repo_type_data`
 * add filter `gu_install_remote_install` for remote install
 * add filter `gu_api_url_type` for API URL data